feat: with re-instantiation
