+++ preprocessing/nextclade/src/loculus_preprocessing/__init__.py
@@ -1,5 +1,6 @@
 from . import backend as backend
 from . import config as config
 from . import datatypes as datatypes
+from . import external_services as external_services
 from . import nextclade as nextclade
 from . import prepro as prepro
