+++ preprocessing/nextclade/src/loculus_preprocessing/datatypes.py
@@ -1,7 +1,10 @@
+import logging
 from collections.abc import Iterable
 from dataclasses import dataclass, field
 from enum import StrEnum, unique
 from typing import Any, Final
+
+logger = logging.getLogger(__name__)
 
 AccessionVersion = str
 GeneName = str
@@ -259,3 +262,13 @@ class MoleculeType(StrEnum):
 class Topology(StrEnum):
     LINEAR = "linear"
     CIRCULAR = "circular"
+
+
+def _internal_error_message(message: str) -> str:
+    full = f"Internal Error. {message} Please contact the administrator."
+    logger.error(full)
+    return full
+
+
+def raw_internal_error(message: str) -> RawProcessingResult:
+    return processing_error(_internal_error_message(message))
