+++ preprocessing/nextclade/tests/test_metadata_processing_functions.py
@@ -1490,7 +1490,7 @@ def test_call_function_converts_raw_errors_to_annotations() -> None:
         "is_insdc_ingest_group": False,
     }
 
-    result = ProcessingFunctions.call_function(
+    result = ProcessingFunctions(config=Config()).call_function(
         "process_options",
         args=args,
         input_data={"input": "NotAnOption"},
